+++ src/private/cli/TestNovaCliBumpConfirmationIsEnabled.ps1
@@ -1,7 +1,0 @@
-function Test-NovaCliBumpConfirmationIsEnabled {
-    [CmdletBinding()]
-    param()
-
-    return $env:NOVA_CLI_CONFIRM_BUMP -eq '1'
-}
-